// src/interfaces/Theme.ts
export interface Colors {
  colorAltHighlight: string;
  colorAltHighlightDark: string;
  colorAltHighlightDarker: string;
  colorAltHighlightDarkest: string;
  colorAltHighlightDarkestS: string;
  colorAltHighlightLight: string;
  colorAltHighlightLighter: string;
  colorAltHighlightLightest: string;
  colorAltHighlightLightestS: string;
  colorAltPrimary: string;
  colorAltPrimaryDark: string;
  colorAltPrimaryDarker: string;
  colorAltPrimaryDarkest: string;
  colorAltPrimaryDarkestS: string;
  colorAltPrimaryLight: string;
  colorAltPrimaryLighter: string;
  colorAltPrimaryLightest: string;
  colorAltPrimaryLightestS: string;
  colorAltSecondary: string;
  colorAltSecondaryDark: string;
  colorAltSecondaryDarker: string;
  colorAltSecondaryDarkest: string;
  colorAltSecondaryDarkestS: string;
  colorAltSecondaryLight: string;
  colorAltSecondaryLighter: string;
  colorAltSecondaryLightest: string;
  colorAltSecondaryLightestS: string;
  colorBlack: string;
  colorComplementary1: string;
  colorComplementary2: string;
  colorComplementary3: string;
  colorComplementary4: string;
  colorComplementary5: string;
  colorDanger: string;
  colorDangerDark: string;
  colorDangerDarker: string;
  colorDangerDarkest: string;
  colorDangerLighest: string;
  colorDangerLight: string;
  colorDangerLighter: string;
  colorGray: string;
  colorGrayDark: string;
  colorGrayDarker: string;
  colorGrayDarkest: string;
  colorGrayLight: string;
  colorGrayLighter: string;
  colorGrayLightest: string;
  colorGrayLightestS: string;
  colorHighlight: string;
  colorHighlightDark: string;
  colorHighlightDarker: string;
  colorHighlightDarkest: string;
  colorHighlightLight: string;
  colorHighlightLighter: string;
  colorHighlightLightest: string;
  colorInfo: string;
  colorInfoDark: string;
  colorInfoDarker: string;
  colorInfoDarkest: string;
  colorInfoLighest: string;
  colorInfoLight: string;
  colorInfoLighter: string;
  colorPrimary: string;
  colorPrimaryDark: string;
  colorPrimaryDarker: string;
  colorPrimaryDarkest: string;
  colorPrimaryLight: string;
  colorPrimaryLighter: string;
  colorPrimaryLightest: string;
  colorSecondary: string;
  colorSecondaryDark: string;
  colorSecondaryDarker: string;
  colorSecondaryDarkest: string;
  colorSecondaryLight: string;
  colorSecondaryLighter: string;
  colorSecondaryLightest: string;
  colorSecondaryLightestS: string;
  colorSuccess: string;
  colorSuccessDark: string;
  colorSuccessDarker: string;
  colorSuccessDarkest: string;
  colorSuccessLighest: string;
  colorSuccessLight: string;
  colorSuccessLighter: string;
  colorWarning: string;
  colorWarningDark: string;
  colorWarningDarker: string;
  colorWarningDarkest: string;
  colorWarningLighest: string;
  colorWarningLight: string;
  colorWarningLighter: string;
  colorWhite: string;
  colorBgHover: string;
}

export interface Variables {
  fontFamily: string;
  appBarHeight: string;
  drawerWidth: string;
  logoWidth: string;
  responsiveLogoWidth: string;
}